Enabled some xfail tests, fixed bugs in tests or in code, and filed issues for seemingly real bugs.

+++ src/laser_generic/births.py
@@ -141,7 +141,7 @@ class Births:
             timing.append(int(delta.total_seconds() * 1_000_000))
         self._metrics.append(timing)
 
-        model.patches.populations[tick + 1, :] += todays_births
+        model.patches.populations[tick + 1, :] += todays_births.astype(np.uint32)
 
         return
